Editor: Move LSP related code from codeeditor.py to a new file
Also, move some the TabSwitcherWidget to its own file
